requirements.txt
appdirs==1.4.4
asgiref==3.2.10
blis==0.7.3
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
ChatterBot==1.0.8
colorama==0.4.4
comtypes==1.1.7
cymem==2.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.2
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
image==1.5.33
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
mathparse==0.1.2
murmurhash==1.0.4
mysql==0.0.2
mysqlclient==2.0.1
numpy==1.19.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
psycopg2==2.8.6
pygame==2.0.0
Pygments==2.7.1
pypiwin32==223
python-dateutil==2.8.1
python-decouple==3.4
pyttsx3==2.90
pytz==2020.1
pywin32==228
qrcode==6.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
srsly==1.0.4
thinc==7.4.3
tqdm==4.54.1
urllib3==1.25.11
virtualenv==20.0.35
wasabi==0.8.0
whitenoise==5.2.0
